feat: implement responsive design and update styles
