+++ lib/ui/widgets/text/display_large_text.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class DisplayLargeText extends StatelessWidget {
+  const DisplayLargeText(this.text, {super.key});
+
+  final String text;
+
+  @override
+  Widget build(BuildContext context) {
+    return Text(
+      text,
+      textAlign: TextAlign.center,
+      style: Theme.of(context).textTheme.displayLarge,
+    );
+  }
+}
