+++ lib/ui/widgets/username.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:portfolio/core/utils/constants.dart';
+import 'package:portfolio/ui/widgets/text/body_medium_text.dart';
 
 class Username extends StatelessWidget {
   const Username({super.key});
@@ -7,19 +8,7 @@ class Username extends StatelessWidget {
   Widget build(BuildContext context) {
     return const Padding(
       padding: EdgeInsets.all(16),
-      child: Text(Constants.username),
-    );
-  }
-}
-
-class Name extends StatelessWidget {
-  const Name({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const Padding(
-      padding: EdgeInsets.all(16),
-      child: Text(Constants.name, textAlign: TextAlign.center),
+      child: BodyMediumText(Constants.username),
     );
   }
 }
